+++ ContentPatcher/Framework/Conditions/ConditionType.cs
@@ -42,8 +42,8 @@ namespace ContentPatcher.Framework.Conditions
         /// <summary>An event ID the player saw.</summary>
         HasSeenEvent,
 
-        /// <summary>A response ID of an answer given by the player.</summary>
-        HasDialogueQuestionAnswered,
+        /// <summary>The response IDs for the player's answers to question dialogues.</summary>
+        HasDialogueAnswer,
 
         /// <summary>The special items in the player's wallet.</summary>
         HasWalletItem,